+++ packages/buefy-next/src/components/autocomplete/Autocomplete.vue
@@ -1,5 +1,9 @@
 <template>
-    <div class="autocomplete control" :class="{ 'is-expanded': expanded }">
+    <div
+        class="autocomplete control"
+        :class="{ 'is-expanded': expanded }"
+        v-bind="rootAttrs"
+    >
         <b-input
             v-model="newValue"
             ref="input"
@@ -15,7 +19,7 @@
             :autocomplete="newAutocomplete"
             :use-html5-validation="false"
             :aria-autocomplete="ariaAutocomplete"
-            v-bind="$attrs"
+            v-bind="inputAttrs"
             @update:model-value="onInput"
             @focus="focused"
             @blur="onBlur"
@@ -107,6 +111,7 @@
 </template>
 
 <script>
+import config from '../../utils/config'
 import {
     getValueByPath,
     removeElement,
@@ -161,7 +166,11 @@ export default {
             default: () => ['Tab', 'Enter']
         },
         selectableHeader: Boolean,
-        selectableFooter: Boolean
+        selectableFooter: Boolean,
+        compatFallthrough: {
+            type: Boolean,
+            default: () => config.defaultCompatFallthrough
+        }
     },
     emits: [
         'active',
@@ -313,6 +322,23 @@ export default {
             return this.iconRightClickable
         },
 
+        rootAttrs() {
+            return this.compatFallthrough
+                ? {
+                    class: this.$attrs.class,
+                    style: this.$attrs.style,
+                    id: this.$attrs.id
+                }
+                : {}
+        },
+        inputAttrs() {
+            if (this.compatFallthrough) {
+                const { style: _1, class: _2, id: _3, ...rest } = this.$attrs
+                return rest
+            } else {
+                return this.$attrs
+            }
+        },
         contentStyle() {
             return {
                 maxHeight: toCssWidth(this.maxHeight)